# backend-docs/swagger.yaml
openapi: 3.0.3
info:
  title: Demo Server API
  version: 1.0.0
  description: HTTP API for creating leads and managing them from the admin area.
servers:
  - url: http://localhost
    description: Local Docker Compose server
tags:
  - name: Health
    description: Service health checks.
  - name: Leads
    description: Public lead submission.
  - name: Admin Auth
    description: Admin authentication and session management.
  - name: Admin Leads
    description: Authenticated lead management endpoints.

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: Service is running.

  /api/lead:
    post:
      tags:
        - Leads
      summary: Create a public lead
      operationId: createLead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadRequest"
            example:
              name: Jane Smith
              title: Website redesign
              description: Need a landing page and admin dashboard.
              contact: jane@example.com
      responses:
        "201":
          description: Lead created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateLeadResponse"
              example:
                is_ok: true
                message: created
                payload:
                  id: 6f43a6db-20b2-46a7-b25e-faa4fc33f6f5
        "400":
          description: Invalid request, invalid values, or database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRequest:
                  value:
                    is_ok: false
                    message: invalid request
                    payload: null
                invalidValues:
                  value:
                    is_ok: false
                    message: invalid values
                    payload: null
                databaseError:
                  value:
                    is_ok: false
                    message: database error
                    payload: null

  /api/admin/login:
    post:
      tags:
        - Admin Auth
      summary: Login as admin
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginAdminRequest"
            example:
              login: admin
              password: admin-password
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokensResponse"
              example:
                is_ok: true
                message: login
                payload:
                  refresh: 5GXsUS10RBC7bCg5kQV2mdcbe8FtZWEN6xlBaxMVAokHCsWjueeX1OJ3qFvhiW41
                  access: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
        "400":
          description: Invalid request, invalid credentials, session error, or validation error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRequest:
                  value:
                    is_ok: false
                    message: invalid request
                    payload: null
                invalidValues:
                  value:
                    is_ok: false
                    message: invalid values
                    payload: null
                loginError:
                  value:
                    is_ok: false
                    message: database login error
                    payload: null
                sessionError:
                  value:
                    is_ok: false
                    message: database session error
                    payload: null
        "500":
          description: JWT generation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                is_ok: false
                message: jwt error
                payload: null

  /api/admin/refresh:
    post:
      tags:
        - Admin Auth
      summary: Refresh admin tokens
      operationId: refreshAdminTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshAdminRequest"
            example:
              refresh: 5GXsUS10RBC7bCg5kQV2mdcbe8FtZWEN6xlBaxMVAokHCsWjueeX1OJ3qFvhiW41
      responses:
        "200":
          description: Tokens refreshed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokensResponse"
              example:
                is_ok: true
                message: refresh
                payload:
                  refresh: L3wYGkJR7V0BSMtbhHEkcCKGrUJHWrR07GRNYAOgWbBTY3JtwZ8pLtyC5o9paJCo
                  access: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.example.signature
        "400":
          description: Invalid request, invalid values, expired token, or session error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRequest:
                  value:
                    is_ok: false
                    message: invalid request
                    payload: null
                invalidValues:
                  value:
                    is_ok: false
                    message: invalid values
                    payload: null
                tokenExpired:
                  value:
                    is_ok: false
                    message: token expired
                    payload: null
                sessionError:
                  value:
                    is_ok: false
                    message: database session error
                    payload: null
        "500":
          description: JWT generation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                is_ok: false
                message: jwt error
                payload: null

  /api/admin/logout:
    post:
      tags:
        - Admin Auth
      summary: Logout admin
      operationId: logoutAdmin
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutAdminRequest"
            example:
              refresh: L3wYGkJR7V0BSMtbhHEkcCKGrUJHWrR07GRNYAOgWbBTY3JtwZ8pLtyC5o9paJCo
      responses:
        "200":
          description: Logout succeeded. Empty request body is also accepted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                is_ok: true
                message: logout
                payload: null
        "400":
          description: Invalid request body or session error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRequest:
                  value:
                    is_ok: false
                    message: invalid request
                    payload: null
                sessionError:
                  value:
                    is_ok: false
                    message: database session error
                    payload: null
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/admin/leads:
    get:
      tags:
        - Admin Leads
      summary: List leads
      operationId: listLeads
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
          description: Filter by exact lead status.
          example: new
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Search text. Checked before query and search aliases.
          example: website
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Search text alias.
          example: website
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search text alias.
          example: website
        - name: date_from
          in: query
          required: false
          schema:
            type: string
          description: Start creation date. Accepts RFC3339 date-time or YYYY-MM-DD.
          examples:
            date:
              value: "2026-05-01"
            dateTime:
              value: "2026-05-01T00:00:00Z"
        - name: date_to
          in: query
          required: false
          schema:
            type: string
          description: End creation date. Accepts RFC3339 date-time or YYYY-MM-DD.
          examples:
            date:
              value: "2026-05-31"
            dateTime:
              value: "2026-05-31T23:59:59Z"
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 20
            maximum: 100
          description: Number of leads to return. Values above 100 are capped by storage.
          example: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of leads to skip.
          example: 0
        - name: sort
          in: query
          required: false
          schema:
            type: string
            enum:
              - client_name
              - clientname
              - name
              - status
              - created_at
              - createdat
              - created
              - updated_at
              - updatedat
              - updated
          description: Sort field. Defaults to created_at.
          example: created_at
        - name: order
          in: query
          required: false
          schema:
            type: string
            enum:
              - ASC
              - DESC
            default: DESC
          description: Sort order. Invalid values fall back to DESC.
          example: DESC
      responses:
        "200":
          description: Lead list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsListResponse"
              example:
                is_ok: true
                message: leads
                payload:
                  items:
                    - id: 6f43a6db-20b2-46a7-b25e-faa4fc33f6f5
                      name: Jane Smith
                      title: Website redesign
                      description: Need a landing page and admin dashboard.
                      contact: jane@example.com
                      status: new
                      created_at: "2026-05-04T10:00:00Z"
                      updated_at: "2026-05-04T10:00:00Z"
                  count: 1
                  limit: 20
                  offset: 0
        "400":
          description: Invalid query or database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidQuery:
                  value:
                    is_ok: false
                    message: invalid query
                    payload: null
                databaseError:
                  value:
                    is_ok: false
                    message: database error
                    payload: null
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/admin/leads/{id}:
    get:
      tags:
        - Admin Leads
      summary: Get lead by ID
      operationId: getLeadById
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      responses:
        "200":
          description: Lead found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadResponse"
              example:
                is_ok: true
                message: lead
                payload:
                  id: 6f43a6db-20b2-46a7-b25e-faa4fc33f6f5
                  name: Jane Smith
                  title: Website redesign
                  description: Need a landing page and admin dashboard.
                  contact: jane@example.com
                  status: new
                  created_at: "2026-05-04T10:00:00Z"
                  updated_at: "2026-05-04T10:00:00Z"
        "400":
          description: Invalid ID or database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidValues:
                  value:
                    is_ok: false
                    message: invalid values
                    payload: null
                databaseError:
                  value:
                    is_ok: false
                    message: database error
                    payload: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /api/admin/leads/{id}/status:
    patch:
      tags:
        - Admin Leads
      summary: Change lead status
      operationId: changeLeadStatus
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/LeadID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeLeadStatusRequest"
            example:
              status: in_progress
      responses:
        "200":
          description: Status changed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                is_ok: true
                message: status changed
                payload: null
        "400":
          description: Invalid request, invalid values, or database error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidRequest:
                  value:
                    is_ok: false
                    message: invalid request
                    payload: null
                invalidValues:
                  value:
                    is_ok: false
                    message: invalid values
                    payload: null
                databaseError:
                  value:
                    is_ok: false
                    message: database error
                    payload: null
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    LeadID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Lead UUID.
      example: 6f43a6db-20b2-46a7-b25e-faa4fc33f6f5

  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            is_ok: false
            message: unauthorized
            payload: null
    NotFound:
      description: Resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            is_ok: false
            message: not found
            payload: null

  schemas:
    ResponseBase:
      type: object
      required:
        - is_ok
        - message
        - payload
      properties:
        is_ok:
          type: boolean
        message:
          type: string
        payload:
          nullable: true

    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseBase"
        - type: object
          properties:
            is_ok:
              type: boolean
              enum:
                - false
            payload:
              nullable: true

    SuccessResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseBase"
        - type: object
          properties:
            is_ok:
              type: boolean
              enum:
                - true
            payload:
              nullable: true

    CreateLeadRequest:
      type: object
      required:
        - name
        - title
        - description
        - contact
      properties:
        name:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        contact:
          type: string
          minLength: 1

    LoginAdminRequest:
      type: object
      required:
        - login
        - password
      properties:
        login:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
          format: password

    RefreshAdminRequest:
      type: object
      required:
        - refresh
      properties:
        refresh:
          type: string
          minLength: 1

    LogoutAdminRequest:
      type: object
      properties:
        refresh:
          type: string
          description: Optional refresh token to revoke.

    ChangeLeadStatusRequest:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          minLength: 1
          maxLength: 32

    ResponseLeadCreate:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid

    ResponseTokens:
      type: object
      required:
        - refresh
        - access
      properties:
        refresh:
          type: string
        access:
          type: string

    ResponseLead:
      type: object
      required:
        - id
        - name
        - title
        - description
        - contact
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        title:
          type: string
        description:
          type: string
        contact:
          type: string
        status:
          type: string
          maxLength: 32
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ResponseLeadsList:
      type: object
      required:
        - items
        - count
        - limit
        - offset
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ResponseLead"
        count:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 0
        offset:
          type: integer
          minimum: 0

    CreateLeadResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseBase"
        - type: object
          properties:
            is_ok:
              type: boolean
              enum:
                - true
            payload:
              $ref: "#/components/schemas/ResponseLeadCreate"

    TokensResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseBase"
        - type: object
          properties:
            is_ok:
              type: boolean
              enum:
                - true
            payload:
              $ref: "#/components/schemas/ResponseTokens"

    LeadResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseBase"
        - type: object
          properties:
            is_ok:
              type: boolean
              enum:
                - true
            payload:
              $ref: "#/components/schemas/ResponseLead"

    LeadsListResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseBase"
        - type: object
          properties:
            is_ok:
              type: boolean
              enum:
                - true
            payload:
              $ref: "#/components/schemas/ResponseLeadsList"